+++ core/src/alchyr/taikoedit/editor/tools/EditorTool.java
@@ -4,7 +4,7 @@ import alchyr.taikoedit.core.layers.EditorLayer;
 import alchyr.taikoedit.editor.views.MapView;
 import alchyr.taikoedit.editor.views.ViewSet;
 import alchyr.taikoedit.editor.maps.EditorBeatmap;
-import alchyr.taikoedit.util.input.MouseHoldObject;
+import alchyr.taikoedit.core.input.MouseHoldObject;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
 
@@ -34,7 +34,7 @@ public abstract class EditorTool {
 
     public abstract boolean supportsView(MapView view);
 
-    public MouseHoldObject click(MapView view, int x, int y, int button, int modifiers)
+    public MouseHoldObject click(MapView view, float x, float y, int button, int modifiers)
     {
         return null;
     }
